autoresearch/pyproject.toml
[project]
name = "autoresearch"
version = "0.2.0"
description = "Autonomous Gemma runtime optimization loop for Apple Silicon"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "psutil>=7.1.0",
    "pyyaml>=6.0.2",
]
